+++ src/atom/product.atom.ts
@@ -1,6 +1,7 @@
 import { atom } from 'recoil';
 import { UnreadCount } from '../App';
-import { MessageType } from '../components/chat/ChatModal';
+
+
 
 // import { selector } from 'recoil';
 
@@ -21,6 +22,56 @@ export const ChatId = atom({
   default: '', // 기본값
 });
 
+export const chatTypingValue = atom({
+  key: 'chatTypingValue', // 고유한 키
+  default: '', // 기본값
+});
+
+export const mainChatModalOpen = atom({
+  key: 'mainChatModalOpen', // 고유한 키
+  default: false, // 기본값
+});
+type productDetails = {
+  image: string;
+  title: string;
+  price: number;
+  id: number;
+};
+type otherUserDetails = {
+  nickname: string;
+  user_img: string;
+};
+
+
+export const otherUserDetail = atom<otherUserDetails | null>({
+  key: 'otherUserDetail', 
+  default: null, // Now null is a valid default value
+});
+
+export const productDetail = atom<productDetails | null>({
+  key: 'productDetail',
+  default: null, // Now null is a valid default value
+});
+
+export type MessageType = {
+  id: number;
+  content: string;
+  author_id: string;
+  chat_id: string;
+  item_id: number;
+  others_id: string;
+  users?: UserType; // 사용자 닉네임을 포함할 수 있는 옵셔널 프로퍼티
+  created_at: number;
+};
+
+export type UserType = {
+  id: string;
+  email: string;
+  lastMessage?: string; // lastMessage 속성 추가 (옵셔널로 처리)
+  nickname: string;
+};
+
+
 export const sendMessages = atom<MessageType[]>({
   key: 'sendMessagesState',
   default: [], // 초기값은 빈 배열
@@ -30,8 +81,6 @@ export const chatRoomsState = atom<ChatRoom[]>({
   key: 'chatRoomsState',
   default: [], // 초기값은 빈 배열
 });
-
-type ProductImage = string | null;
 
 export type ChatRoom = {
   author_id: string;
@@ -75,6 +124,50 @@ export const UnreadCounts = atom<UnreadCount[]>({
   key: 'UnreadCounts',
   default: [],
 });
+
+
+export const updateMesaages = atom<MessagePayload>({
+  key: 'updateMesaages',
+  default:{
+    commit_timestamp: '',
+    errors: null , // Replace 'any' with a more specific type if possible
+    eventType: '',
+    new: {
+      author_id: '',
+      chat_id: '',
+      content: '',
+      created_at: '',
+      id: 0,
+      item_id: 0,
+      others_id: '',
+    },
+    
+    old: '', // Replace 'any' with a more specific type if possible
+    schema: '',
+    table: '',
+  }
+   
+});
+export type MessagePayload = {
+  commit_timestamp: string;
+  errors: null | any; // Replace 'any' with a more specific type if possible
+  eventType: string;
+  new: {
+    author_id: string;
+    chat_id: string;
+    content: string;
+    created_at: string;
+    id: number;
+    item_id: number;
+    others_id: string | null;
+  };
+  
+  old?: any; // Replace 'any' with a more specific type if possible
+  schema: string;
+  table: string;
+};
+
+
 
 // // ======================================
 // // 사용법
